requirements.txt
pandas
elasticsearch
python-dotenv